requirements.txt
plotly.express
pandas
sys
pytest